requirements.txt
flask
flask_jwt_extended
pymysql
cryptography
pycryptodome
redis
firebase_admin